requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# LLM Providers
langchain-google-genai>=0.0.5
langchain-openai>=0.0.5

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Tools
wikipedia>=1.4.0
duckduckgo-search>=4.0.0

# Optional: For better async support
aiohttp>=3.9.0

# Optional: For API serving
fastapi>=0.104.0
uvicorn>=0.24.0

# Optional: For testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
